requirements.txt
aniso8601==9.0.1
attrs==23.2.0
behave==1.2.6
black==23.12.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
install==1.3.5
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
parse==1.20.0
parse-type==0.6.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
